Added API functionality to the reservation page.
Script was not loading from raw.github so I had to download the 'API'
and implement into the application itself. Updated the initialize times
as well as the updateTimes functions so that the available times are
determined by the date selected.

Added the submitAPI functionality to the handleSubmit function. This
will also console log the return result as a boolean.